+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# vNext
+
+* Fixes #5: Passing holes to createPolygon results in double free
+
 # v0.1.7.1
 
 * Removes problematic helpers.h